+++ include/vectorSurface.hpp
@@ -8,52 +8,52 @@
 NAMESPACESTART
 class VectorSurface : public VectorAnalysis3D{
   private:
-    std::function<float(float, float)> xFunc;
-    std::function<float(float, float)> yFunc;
-    std::function<float(float, float)> zFunc;
+    std::function<double(double, double)> xFunc;
+    std::function<double(double, double)> yFunc;
+    std::function<double(double, double)> zFunc;
 
     std::vector<Dmath::Vec3D> mainSurface;
     std::vector<Dmath::Vec3D> surfaceNormals;
 
-    float epsilon = 0.001;
+    double epsilon = 0.001;
 
-    Dmath::Vec3D calculatePartialDerivativeU(float u, float v);
-    Dmath::Vec3D calculatePartialDerivativeV(float u, float v);
+    Dmath::Vec3D calculatePartialDerivativeU(double u, double v);
+    Dmath::Vec3D calculatePartialDerivativeV(double u, double v);
 
     void createVectorSurface();
 
-    VectorSurface(std::function<float(float, float)> xFunc, std::function<float(float, float)> yFunc, std::function<float(float, float)> zFunc, float systemStart, float systemStopp, float resolution);
+    VectorSurface(std::function<double(double, double)> xFunc, std::function<double(double, double)> yFunc, std::function<double(double, double)> zFunc, double systemStart, double systemStopp, double resolution);
 
 
-    float integrateSurfaceArea();
+    double integrateSurfaceArea();
 
   public:
-    static VectorSurface createStandardSurface(std::function<float(float, float)> xFunc, std::function<float(float, float)> yFunc, std::function<float(float, float)> zFunc);
+    static VectorSurface createStandardSurface(std::function<double(double, double)> xFunc, std::function<double(double, double)> yFunc, std::function<double(double, double)> zFunc);
 
-    static VectorSurface createCustomSurface(std::function<float(float, float)> xFunc, std::function<float(float, float)> yFunc,  std::function<float(float, float)> zFunc, float systemStart, float systemStopp, float resolution);
+    static VectorSurface createCustomSurface(std::function<double(double, double)> xFunc, std::function<double(double, double)> yFunc,  std::function<double(double, double)> zFunc, double systemStart, double systemStopp, double resolution);
 
     void calculateSurfaceNormals();
 
-    inline void setEpsilon(float epsilon) { this->epsilon = epsilon;       }
-    inline float getEpsilon()             { return this->epsilon;          }
-    inline float getSurfaceArea()         { return integrateSurfaceArea(); }
+    inline void setEpsilon(double epsilon) { this->epsilon = epsilon;       }
+    inline double getEpsilon()             { return this->epsilon;          }
+    inline double getSurfaceArea()         { return integrateSurfaceArea(); }
 
-    inline std::function<float(float,float)> getXFunction() {return this->xFunc;}
-    inline std::function<float(float,float)> getYFunction() {return this->yFunc;}
-    inline std::function<float(float,float)> getZFunction() {return this->zFunc;}
+    inline std::function<double(double,double)> getXFunction() {return this->xFunc;}
+    inline std::function<double(double,double)> getYFunction() {return this->yFunc;}
+    inline std::function<double(double,double)> getZFunction() {return this->zFunc;}
 
     inline Dmath::Vec3D getSurfaceNormales(size_t index){return this->surfaceNormals[index];}
 
-    Dmath::Vec3D getVectorAtPointT(float t);
+    Dmath::Vec3D getVectorAtPointT(double t);
     
-    float maxX();
-    float minX();
-    float maxY();
-    float minY();
-    float maxZ();
-    float minZ();
+    double maxX();
+    double minX();
+    double maxY();
+    double minY();
+    double maxZ();
+    double minZ();
 
-    float calculatePerimeter();
+    double calculatePerimeter();
 
 };
